refactor: extract ConnectionManager with disposal_guard and Drop (#165)
Move all WebSocket background tasks (connect, reconnect, watchdog,
token-expiry) into ConnectionManager struct with RAII cleanup.
Replaces 4 on_cleanup calls with a single Drop implementation.
